RemoteScrollingCoordinatorProxy needs to receive NativeWebWheelEvents https://bugs.webkit.org/show_bug.cgi?id=252860 rdar://105852101
Reviewed by Sam Weinig.

RemoteLayerTreeEventDispatcher will get a MomentumEventDispatcher soon, and MomentumEventDispatcher
needs to be fed a ScrollingAccelerationCurve which is derived from a native event, so these
native events need to get to RemoteScrollingCoordinatorProxy.

So change the argument to handleWheelEvent() to be a NativeWebWheelEvent.

* Source/WebKit/UIProcess/RemoteLayerTree/RemoteScrollingCoordinatorProxy.cpp:
(WebKit::RemoteScrollingCoordinatorProxy::handleWheelEvent):
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteScrollingCoordinatorProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::handleWheelEvent):

Canonical link: https://commits.webkit.org/260784@main