Reduce Bridging and Runtime Creation of Strings https://bugs.webkit.org/show_bug.cgi?id=251687
Reviewed by NOBODY (OOPS!).

In many places, we needlessly bridge for the sake of calling
Objective-C methods and Core Foundation functions where the native
functions/methods work just fine. This PR replaces the bridging with
calling the functions associated with the variable type. This is focused
primarily on strings but some fixes required changing other variables to
CF types or Objective-C NS types.

*Source\WTF\wtf\cocoa\FileSystemCocoa.mm:
*Source\WebCore\bridge\objc\objc_class.mm:
*Source\WebCore\platform\cocoa\MIMETypeRegistryCocoa.mm:
*Source\WebCore\platform\gamepad\cocoa\GameControllerGamepadProvider.mm:
*Source\WebKit\UIProcess\API\Cocoa\WKWebView.mm:
*Source\WebKit\UIProcess\API\Cocoa\_WKSystemPreferences.mm:
*Source\WebKit\UIProcess\API\Cocoa\_WKSystemPreferencesInternal.h:
*Source\WebKit\UIProcess\Cocoa\WebProcessPoolCocoa.mm:
*Source\WebKit\UIProcess\Inspector\mac\WebInspectorUIProxyMac.mm:
*Source\WebKit\UIProcess\WebsiteData\Cocoa\WebsiteDataStoreCocoa.mm:
*Source\WebKit\UIProcess\mac\WKPrintingView.mm:
*Source\WebKitLegacy\mac\Misc\WebNSDataExtras.mm:(_web_capitalizeRFC822HeaderFieldName):
This function was heavily rewritten to call Objective-C functions with
minimal bridging. In addition, unlike the way the function was written
previously, this function frees the backing buffer. However, as a
caveat, and because we cannot change this, said buffer MUST use
malloc/free instead of the standard and established way we allocate and
free temporary buffers.