AX: Move the AttributedString for TextMarkerRange functionality out of the wrapper into the AXCoreObject implementation so that it can be cached in isolated tree mode. https://bugs.webkit.org/show_bug.cgi?id=252193 <rdar://problem/105416277>
Reviewed by Chris Fleizach.

NSAttrbutedString for a given TextMarkerRange functionality was implemented in the AX object platform wrapper code. There was code duplication in the Mac and IOS implementations and some outdated code. this patch does the following:
- Moves this functionality implementation out of the wrappers into the AXObject implementation.
- Reuses some common code for Mac and iOS in AccessibilityObjectCocoa.mm.
- Keeps the platform specific code in AccessibilityObjectMac and IOS.mm respectively.
- Lots of code cleanup and strengthening.
There is a follow up patch that will move most of the AttributedString creation and modification functions into an AXAttributedStringBuilder class to be used in both Mac and iOS.
This is a prequel to a change to cache the AttributedString for some types of AX objects in ITM and therfore reduce main thread hits to serve these client requests.

Canonical link: https://commits.webkit.org/260800@main