[CMake] Rework decision to enable -gsplit-dwarf by default https://bugs.webkit.org/show_bug.cgi?id=252679
Reviewed by Adrian Perez de Castro.

-gsplit-dwarf breaks RPM builds due to zero-sized debug files. I could
override this with -DDEBUG_FISSION=OFF, but we're probably trying to be
_too_ helpful here by injecting debug flags into non-debug build types.
I doubt Fedora is the only distro that's not prepared for -gsplit-dwarf
so let's only inject it if the CMake build type is a debug build type,
and if ENABLE_DEVELOPER_MODE (set e.g. by build-webkit) which is a good
indication that incremental builds are desired. (The primary goal of
-gsplit-dwarf is to make incremental builds faster.) That way, we
improve CMake's default behavior for developers doing incremental
builds, but anyone manually passing CFLAGS will still get exactly what
they request without modification, and we avoid breaking find-debuginfo
scripts or other debuginfo machinery that CMake does not know about.

* Source/cmake/OptionsCommon.cmake:

Canonical link: https://commits.webkit.org/260818@main