SourceBuffer.timestampOffset usage doesn't behave correctly with webm content. https://bugs.webkit.org/show_bug.cgi?id=252782 rdar://105801920
Reviewed by Youenn Fablet.

The libwebmparser enforced checking the size of the WebM Segment parent.
When we append a cluster a second time, it would have made the total
size of the webm larger than the parent Segment could fit and the parser
aborted with an error. That error wasn't propagated to JS.

Considering a webm is only ever made of one Segment, and that the Matroska
specification allows for an element size to be unknown and other browsers
ignore that size, the simplest way to get around is to ignore the
Segment's size and assume it is always unknown.

This allows to add as many cluster as one wants, regardless of the init
segment added.
libwebm is a 3rd party library, however we have already modified it to
suits our requirement.

* LayoutTests/media/media-source/content/test-opus-manifest.json:
* LayoutTests/media/media-source/content/test-opus.webm: Remux the file with ffmpeg so it reproduces the same issue as the vorbis one.
* LayoutTests/media/media-source/media-source-webm-fragment-expected.txt: Added.
* LayoutTests/media/media-source/media-source-webm-fragment.html: Added.
* Source/ThirdParty/libwebrtc/Source/third_party/libwebm/webm_parser/src/webm_parser.cc:

Canonical link: https://commits.webkit.org/260822@main