Remove old and deprecated UIProcess functions https://bugs.webkit.org/show_bug.cgi?id=251879
Reviewed by Alex Christensen.

The code in question is literally code that was commented to be removed
under certain conditions that have been met years ago, such as no longer
compiling for Safari 8 and when Mavericks support is gone.

This is very very old code that is the removal of which is obvious but
has yet to be done.

Canonical link: https://commits.webkit.org/260831@main