Reduce build times by refactoring ElementIterator.h and friends https://bugs.webkit.org/show_bug.cgi?id=252107
Reviewed by Simon Fraser.

ElementIterator.h is a common cause of ElementInlines.h being unnecessarily included. A common
strategy to mitigate this kind of header include pattern is to move the implementations of inlined
methods into an ElementIteratorInlines.h file, and include ElementInlines.h from that file instead.
This allows ElementIterator.h to be included by files who need to reference the type itself, but
don't need to call methods of that type. (Or at least, not the inlined methods).

Howevwer, this spawned a hydra-like situation where moving inlined methods into an Inlines.h file
caused downstream headers with their own inlines to pull in this new Inlines.h file, which is
mitigated by moving those inlined methods into their own Inlines.h file, and so on ad nauseum.

Build times are improved minorly by this change; the number of times ElementInlines.h is included in
a Unified Build of WebCore dropped from 220 to 180, the cost of including ElementIterator.h dropped
by about 80%, the number of times it is included dropped from 113 to 106, but many files (about 90)
do end up needing to include various *IteratorInlines.h headers, which limits the savings to an extent.

Canonical link: https://commits.webkit.org/260856@main