[GTK] Switch back to 0 painting threads https://bugs.webkit.org/show_bug.cgi?id=253108
Reviewed by NOBODY (OOPS!).

We thought it might be safe to use 1 painting thread, but this does not
seem to be true, so go back to 0 for now.

I've also removed the DEVELOPER_MODE and platform preprocessor guards
because there's not really any need for them. It's fine to gate the
functionality behind an environment variable. The implementation is
compiled unconditionally anyway.

Lastly, change what happens when an invalid value is passed via the
environment variable. Currently the "default value" 4 is used only when
you pass an invalid value via the environment variable. That's pretty
strange -- it's "default" only in the sense of default to use when an
invalid value is selected -- so let's switch to using the actual default
value 0.

* Source/WebCore/platform/graphics/nicosia/NicosiaPaintingEngine.cpp:
(Nicosia::PaintingEngine::create):