[UI-side compositing] Assertion under RemoteLayerTreeEventDispatcher::displayLink() running tests https://bugs.webkit.org/show_bug.cgi?id=253112 rdar://106051454
Reviewed by Tim Horton.

RemoteScrollingCoordinatorProxy owns a RemoteLayerTreeEventDispatcher which needs to get back to
the DrawingAreaProxy to cancel a DisplayLink on cleanup. This means that the RemoteScrollingCoordinatorProxy
needs to be destroyed before the DrawingAreaProxy, so ensure that in WebPageProxy.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::setDrawingArea):
(WebKit::WebPageProxy::resetState): Call setDrawingArea(nullptr) so we run the same cleanup code.

Canonical link: https://commits.webkit.org/260975@main