[WGSL] Add a tiny DSL for declaring built-in types https://bugs.webkit.org/show_bug.cgi?id=252251 rdar://105457268
Reviewed by Myles C. Maxfield.

Initializing all the built-in types in C++ is quite verbose, so this patch adds a tiny
embedded DSL in ruby for declaring such types (similar to JavaScriptCore/BytecodeList.rb)

NOTE: We previously had a PR with this change (#10099) but decided to use a JavaScript
DSL instead. That caused some build issues, so we're going back to Ruby.

* Source/WebGPU/WGSL/Overload.h:
* Source/WebGPU/WGSL/TypeCheck.cpp:
(WGSL::TypeChecker::TypeChecker):
* Source/WebGPU/WGSL/TypeDeclarations.rb: Added.
* Source/WebGPU/WGSL/generator/main.rb: Added.
* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/260991@main