ImageBufferIOSurfaceBackend should use bitmap CGImages when drawing to bitmap context https://bugs.webkit.org/show_bug.cgi?id=252986 rdar://problem/105971637
Reviewed by NOBODY (OOPS!).

Previously when drawing IOSurface -based ImageBuffer to a bitmap CGContext,
the IOSurface would be wrapped into a CGImage and drawn. CG would do
the readback, construct a bitmap image and draw it as a bitmap and store the
bitmap to the CGImage. As a fix up step, we would update the
CGIOSurfaceContext artificially to destroy the CGImage, so that it would
not hold the bitmap cache.

Do this explicitly by providing a reference to the IOSurface via bitmap
image. This codepath is only used for drawing and only on specific
backends. These limitations make us be sure that the image does not get
long term references, and as such future modifications do not change
the already rendered content.

 This has the benefit of:
 - Being simpler to reason about: no cleanup step.
 - Image being a reference omits a copy from the IOSurface buffer,
   and thus speeds up the draw considerably.
 - Having the handle to the bitmap image enables further small
   optimizations in future patches.

* Source/WebCore/platform/graphics/ImageBuffer.cpp:
(WebCore::ImageBuffer::copyNativeImageForDrawing const):
(WebCore::ImageBuffer::draw):
* Source/WebCore/platform/graphics/ImageBuffer.h:
* Source/WebCore/platform/graphics/ImageBufferBackend.cpp:
(WebCore::ImageBufferBackend::copyNativeImageForDrawing):
* Source/WebCore/platform/graphics/ImageBufferBackend.h:
* Source/WebCore/platform/graphics/cg/ImageBufferIOSurfaceBackend.cpp:
(WebCore::createIOSurfaceBitmapImageReference):
(WebCore::ImageBufferIOSurfaceBackend::copyNativeImageForDrawing):
(WebCore::ImageBufferIOSurfaceBackend::finalizeDrawIntoContext): Deleted.
* Source/WebCore/platform/graphics/cg/ImageBufferIOSurfaceBackend.h:
* Source/WebCore/platform/graphics/cocoa/IOSurface.h:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp:
(WebKit::RemoteImageBufferProxy::copyNativeImageForDrawing const):
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.h: