[UI-side compositing] Throttle AdjustTransientZoom IPC calls https://bugs.webkit.org/show_bug.cgi?id=252365 rdar://105806821
Reviewed by NOBODY (OOPS!).

Throttle calls to AdjustTransientZoom by waiting for a layer tree commit from the web
process after each call to AdjustTransientZoom. This solution currently doesn't work
in this configuration as we only receive layer tree commits when prepopulateRectForZoom
actually causes tiles to get populated. We should either find a way to trigger a
commit for every call to AdjustTransientZoom or wait to land this once we are updating
tiles in AdjustTransientZoom.

* Source/WebKit/UIProcess/RemoteLayerTree/mac/RemoteLayerTreeDrawingAreaProxyMac.h:
* Source/WebKit/UIProcess/RemoteLayerTree/mac/RemoteLayerTreeDrawingAreaProxyMac.mm:
(WebKit::RemoteLayerTreeDrawingAreaProxyMac::didCommitLayerTree):
(WebKit::RemoteLayerTreeDrawingAreaProxyMac::adjustTransientZoom):