[JSC] x64 CCall returnValueGPR is not in m_validGPRs https://bugs.webkit.org/show_bug.cgi?id=253227 rdar://106127760
Reviewed by Mark Lam.

x64's returnValueGPR is not in m_validGPRs. So we cannot bind it to Location.
We should move it to argumentGPR0 if returnValueGPR is not argumentGPR0, this is kind of a hack and we should
change emitCCall in the future to make it more barebone like DFG's callOperation.

* Source/JavaScriptCore/wasm/WasmBBQJIT.cpp:
(JSC::Wasm::BBQJIT::emitCCall):

Canonical link: https://commits.webkit.org/261048@main